services/organizer/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
tmdbv3api==1.9.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
